Feature-gate approx trait implementations
