+++ src/arraytraits.rs
@@ -113,92 +113,6 @@ impl<S, D> Eq for ArrayBase<S, D>
           S::Elem: Eq,
 { }
 
-impl<A, S, D> approx::AbsDiffEq for ArrayBase<S, D>
-where
-    A: approx::AbsDiffEq,
-    A::Epsilon: Clone,
-    S: Data<Elem = A>,
-    D: Dimension,
-{
-    type Epsilon = A::Epsilon;
-
-    fn default_epsilon() -> A::Epsilon {
-        A::default_epsilon()
-    }
-
-    fn abs_diff_eq(&self, other: &ArrayBase<S, D>, epsilon: A::Epsilon) -> bool {
-        if self.shape() != other.shape() {
-            return false;
-        }
-        Zip::from(self)
-            .and(other)
-            .fold_while(true, |_, a, b| {
-                if A::abs_diff_ne(a, b, epsilon.clone()) {
-                    FoldWhile::Done(false)
-                } else {
-                    FoldWhile::Continue(true)
-                }
-            })
-            .into_inner()
-    }
-}
-
-impl<A, S, D> approx::RelativeEq for ArrayBase<S, D>
-where
-    A: approx::RelativeEq,
-    A::Epsilon: Clone,
-    S: Data<Elem = A>,
-    D: Dimension,
-{
-    fn default_max_relative() -> A::Epsilon {
-        A::default_max_relative()
-    }
-
-    fn relative_eq(&self, other: &ArrayBase<S, D>, epsilon: A::Epsilon, max_relative: A::Epsilon) -> bool {
-        if self.shape() != other.shape() {
-            return false;
-        }
-        Zip::from(self)
-            .and(other)
-            .fold_while(true, |_, a, b| {
-                if A::relative_ne(a, b, epsilon.clone(), max_relative.clone()) {
-                    FoldWhile::Done(false)
-                } else {
-                    FoldWhile::Continue(true)
-                }
-            })
-            .into_inner()
-    }
-}
-
-impl<A, S, D> approx::UlpsEq for ArrayBase<S, D>
-where
-    A: approx::UlpsEq,
-    A::Epsilon: Clone,
-    S: Data<Elem = A>,
-    D: Dimension,
-{
-    fn default_max_ulps() -> u32 {
-        A::default_max_ulps()
-    }
-
-    fn ulps_eq(&self, other: &ArrayBase<S, D>, epsilon: A::Epsilon, max_ulps: u32) -> bool {
-        if self.shape() != other.shape() {
-            return false;
-        }
-        Zip::from(self)
-            .and(other)
-            .fold_while(true, |_, a, b| {
-                if A::ulps_ne(a, b, epsilon.clone(), max_ulps) {
-                    FoldWhile::Done(false)
-                } else {
-                    FoldWhile::Continue(true)
-                }
-            })
-            .into_inner()
-    }
-}
-
 impl<A, S> FromIterator<A> for ArrayBase<S, Ix1>
     where S: DataOwned<Elem=A>
 {
